Interface/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.0
openai==1.57.2
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlanalyzer==0.6.0
sqlparse==0.3.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
